+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 # 0.2.0+ - [Unreleased]
 ## Added
 - Automatically maintain a catalog file
+- `./dnstemple.py` to simplify testing the current version
 
 ## Fixed